Refactored CreatePost interface adapters to ClubCreatePost for consistency
