+++ src/main/java/app/AppBuilder.java
@@ -5,7 +5,7 @@ import entity.user.ClubUserFactory;
 import entity.user.StudentUserFactory;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.club_logged_in.ClubLoggedInViewModel;
-import interface_adapter.club_logged_in.create_post.CreatePostViewModel;
+import interface_adapter.club_logged_in.club_create_post.CreatePostViewModel;
 import interface_adapter.login.LoginViewModel;
 import interface_adapter.login.club_login.ClubLoginController;
 import interface_adapter.login.club_login.ClubLoginPresenter;
