+++ src/main/java/interface_adapter/club_logged_in/club_create_post/ClubCreatePostController.java
@@ -1,4 +1,4 @@
-package interface_adapter.club_create_post;
+package interface_adapter.club_logged_in.club_create_post;
 
 import use_case.club_create_post.ClubCreatePostInputBoundary;
 import use_case.club_create_post.ClubCreatePostInputData;
