+++ src/main/java/view/ClubLoggedInView.java
@@ -4,7 +4,7 @@ import javax.swing.*;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
 
-import interface_adapter.club_create_post.ClubCreatePostController;
+import interface_adapter.club_logged_in.club_create_post.ClubCreatePostController;
 import interface_adapter.club_get_posts.ClubGetPostsController;
 import interface_adapter.club_logged_in.ClubLoggedInState;
 import interface_adapter.club_logged_in.ClubLoggedInViewModel;
